% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addins}
\alias{addins}
\title{Tinsel RStudio Addins}
\description{
tinsel provides two RStudio addins to make working with decorators more
intuitive. The first "Source Active File Decoratees" will source decorated
functions from the current file in your RStudio session. The second "Run
Decoratees Selected Line(s)" behaves like RStudio's run lines command.
Decoratored functions within the highlighted section are parsed into the
global environment.
}

