% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{decorators}
\alias{decorators}
\alias{original}
\title{Get Function Decorators or Original Function}
\usage{
decorators(f)

original(f)
}
\arguments{
\item{f}{A decorated function.}
}
\description{
Get the decorators of a function or the original decoratee function from a
decorated function object.
}
\examples{
source_decoratees(tinsel_example('attributes.R'))

# sourced from the 'attributes.R' example file
selector1

# get a list of decorators wrapping a function
decorators(selector1)

# get the original decoratee function of the
# decorated `selector1` function
original(selector1)

}

