% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorated.R
\name{is.decorated}
\alias{is.decorated}
\title{Decorated Functions}
\usage{
is.decorated(f)
}
\arguments{
\item{f}{A function.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Returns \code{TRUE} if the function \code{f} is decorated, otherwise
\code{FALSE}.
}
\examples{
source_decoratees(tinsel_example('timer.R'))

# sourced from the timer.R example file
is.decorated(waldo)
is.decorated(jack)

# it's a function, but not decorated
is.decorated(mean)

# far from the mark
is.decorated(3030)

}

