% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present.R
\name{presents}
\alias{presents}
\title{Load Decorated Functions}
\usage{
presents(file, envir = as.environment(-1), verbose = TRUE)
}
\description{
Given an \code{R} file, the presents function parses and loads decorated
functions into the calling environment (or specified environment).
}

