% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorated.R
\name{print.decorated}
\alias{print.decorated}
\title{Print Decorated Functions}
\usage{
\method{print}{decorated}(x, ...)
}
\arguments{
\item{x}{An \R object.}

\item{\ldots}{Additional arguments passed on to methods.}
}
\description{
A wrapper on top of the standard \code{\link{print.function}} to help make
better sense of decorated functions. Possible arguments are those arguments
which are used by the decoratee, but are not explicitly listed as part of the
final transformed funciton.
}

