% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-decoratees.R
\name{source_decoratees}
\alias{source_decoratees}
\title{Load Decorated Functions}
\usage{
source_decoratees(file)
}
\arguments{
\item{file}{A character string specifying a file path.}

\item{envir}{The \code{\link{environment}} where parsed functions are
assigned.}

\item{verbose}{Defaults to \code{FALSE}, if \code{TRUE} print progress
messages.}
}
\description{
Given an \code{R} file, this function parses and loads decorated functions
into a specified environment.
}
\details{
If any decorators are undefined or if a function is not defined properly
none of the decorated functions parsed are loaded.
}

