% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinsel.R
\docType{package}
\name{tinsel-package}
\alias{tinsel}
\alias{tinsel-package}
\title{Decorators for R Functions}
\description{
Instead of nesting function calls, annotate and transform
functions using "#." comments.
}

