% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{tinsel_example}
\alias{tinsel_example}
\alias{tinsel_examples}
\title{Get Path of a Tinsel Example File}
\usage{
tinsel_example(path)

tinsel_examples()
}
\arguments{
\item{path}{Name of the example file.}
}
\description{
\code{tinsel_example} simplifies getting and returns the system path of an
example file included in the tinsel package. To list the names of all example
files use \code{tinsel_examples}.
}
\examples{
# list all example files
tinsel_examples()

# get the path of a specific example
tinsel_example('attributes.R')
}
\keyword{internal}

