% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{decorators}
\alias{decorators}
\alias{original}
\title{Get Function Decorators or Original Function}
\usage{
decorators(f)

original(f)
}
\arguments{
\item{f}{A decorated function.}
}
\description{
These functions allow programmers to access the decorators of a function or
the original decoratee function.
}

