% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.decorated}
\alias{is.decorated}
\title{Decorated Functions}
\usage{
is.decorated(f)
}
\arguments{
\item{f}{A function.}
}
\value{
Returns \code{TRUE} or \code{FALSE} depending on if the object \code{f} is
decorated.
}
\description{
Check if a function is decorated.
}

