% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print.decorated}
\alias{print.decorated}
\title{Print a Decorated Function}
\usage{
\method{print}{decorated}(x, ...)
}
\arguments{
\item{x}{A decorated function.}

\item{\ldots}{Additional arguments for next print method.}
}
\description{
The \code{print.decorated} function naively prints \code{x} as a function. In
reality, the function printed may be the final of any number of decorators to
a decoratee. To get the original function or the decorators wrapping it use
\code{\link{original}} and \code{\link{decorators}}.
}

